synchronize the code

+++ chain/nuls-chain/src/main/java/io/nuls/chain/info/CmRuntimeInfo.java
@@ -8,7 +8,9 @@ package io.nuls.chain.info;
 public class CmRuntimeInfo {
     public static String dataPath;
 
-    public static String getAssetKey(short chainId, short assetId) {
+    public static String getAssetKey(short chainId, long assetId) {
         return chainId + "-" + assetId;
     }
+    //short i=32767;
+    //public static short next
 }
